Fix dijkstra implementation
Technically not an issue with the dijkstra algorithm itself,
but the function dijkstra would iterate over all exit blocks
with the intent of finding the shortest path to an exit block.
The function exits with an error string if no path can be
found from the starting node to all exit nodes. This prevents
it finding a shortest path in the case that only some exit nodes
are reachable from the starting node, which is what we want.