Create testing pipeline for fleshing
This is for testing if fleshing is working correctly. It uses
the Vulkan CTS as a source of asm files, converts the asm
files to xml files, fleshes the xml files and then executes
the resulting amber files. It is assumed that fleshing is
correct if there are no amber errors.

NOTE: At the time of writing there is an error found when
executing the amber files with Intel MESA drivers due to
a bug in the driver (https://gitlab.freedesktop.org/mesa/mesa/-/issues/6246)