Cross compile SPIR-V from amber files into various target languages
The languages supported are hlsl, glsl and msl. The conversion process uses
spirv-cross to cross compile the SPIR-V from amber files into the target language.
The results are stored in files the appropriate suffix for the target language.
For example, converting to glsl will create .comp files. Naga may be supported
in the future as an alternative to spirv-cross.

Currently, no compilation is done on the target language files. This should
be supported in the future.